Set basic environment and set html file
The webpack, the loaders, the linter and babel were set before diving
into the creation of the app. Once these elements were added to the
project, in the html file a background image was set along with the
container for the data coming from the openWeather API, which will be
added in upcoming commits.