Create an object FORM that oversees the collection of data
In the javascript file a form object was created to hear for the event
"submit" of the form, which would collect the city entered in the input
and make the API request to get all the information available of the
place. The fetch() function was used to get the data, which was
processed with two THEN methods.

The data was added to an array, which would the the promise value to
return. The data consists of the place's name, temperature,
weather's description, latitude and longitude.