Set TouchableOpacity component and localCity state
The localCity state is used to get the city typed by the user. The
final value of this state is used to update the global city state in the
App.js component. Likewise, the TextInput component was populated with
the respective props to update the typed value. Finally, the
TouchableOpacity component wraps the icon to indicate to users it must
be clicked to fetch the data.