Add unitTemp value to object created after fetching data.
Previously the unitTemp state was passed as a prop to all componentes
that required it. However, this led to the units being rendered first
and then the corresponding value. This was solved by adding the value of
the state in the object created from the data given by the API.