Set up Search component
This component will fetch the data from the API once users have entered
a valid city. Likewise, this button was provided with a button to delete
the entered city at once. The delete button will be displayed only when
users write at least one letter, otherwise it will be hidden.