Create lastValidCity state
This state helps to fetch the last valid city data when changing the
metrics. When the current city is not valid then choosing Celsius or
Farenheit would not update the data, but storing the last valid city
would enable this since the api function would receive the city which
data is displayed currently.